// frontend/pharmacy-app/src/enums/pocAssignee.ts
export enum PocAssignee {
  TANMAY = "YXQ3J2iSlDWfuIai7pgUqtQoeHh2",
  AAMER = "dmL0pWAg2eT7Lkqd8V2QvxFSYu83",
  PRINCE = "oeok2cmtvmbBmbcHashUzwkqR4Z2",
  MONICA = "boDk56xO77aijm8ehSqKayejcCo2",
  SAGAR = "tKFCHGbakkZ3Tprgvc1gfoXmT6x2",
  CLARINA = "9XKa3tLEYbUdkn8IA7vgUwi6GJq1",
  PRIYANKA = "XY1QysjJV9fbRJXsRutF4dp0VEz2",
  SANA = "bgq15Rk0zqaaKCznER38jZwIFIE2",
  AKSHITH = "a1wssbmKEBPoiWqtUW3HsVQWCvu2",
}

export const PocAssigneeLabels: Record<PocAssignee, string> = {
  [PocAssignee.TANMAY]: "Tanmay",
  [PocAssignee.AAMER]: "Aamer",
  [PocAssignee.PRINCE]: "Prince",
  [PocAssignee.MONICA]: "Monica",
  [PocAssignee.SAGAR]: "Sagar",
  [PocAssignee.CLARINA]: "Clarina",
  [PocAssignee.PRIYANKA]: "Priyanka",
  [PocAssignee.SANA]: "Sana",
  [PocAssignee.AKSHITH]: "Akshith",
};